15 いいね機能

+++ database/migrations/2020_10_03_065557_create_prefectures_table.php
@@ -15,7 +15,7 @@ class CreatePrefecturesTable extends Migration
     {
         Schema::create('prefectures', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('region_id')->unsigned();
+            $table->bigInteger('region_id')->unsigned();
             $table->integer('code');
             $table->string('name');
             $table->foreign('region_id')->references('id')->on('regions');
